Implement basic reading and writing functions
The genset documentation provides 4 different protocols. The one I was
trying to implement here was the "Protocol for fixed RPM gensets –all
type (SQ and SC, single, dual, three phase)". This is why the fixed_RPM
name is used. Later I learned that this is not the protocol that will use,
but it shouldn't be a problem because the methods
currently implemented should be the same for all protocols (I haven't
implemented the protocol-specific methods for the commands yet). Of
course the names of the classes and namespace should change in the
future and the constant values should be modified to suit the correct
protocol.