Consider a frame valid regardless of command
I realized it shouldn't be necessary to wait for arbitrary commands
(such as GeneratorState or RunTimeState) to arrive before sending a
control command. The documentation just says that the panel will send
its message after it has received a valid message, so just checking that
the frame is valid and that the addresses are correct should be enough.